Nodes can be portals
Portal is a node that may be linked to any other node anywhere in the
playground, not just an adjacent node. Portals are bi-directional and
only ever have two ends.

In the playground file, portal entrance is any character that isn't a
wall, a starting position or a regular node.