PathTracker path finding no longer limited in # of intermediate nodes
We can now use any number of intermediate nodes through which to look for paths. However, tread carefully, as this is an NP-complete problem and, as such, the execution time grows dramatically with additional nodes.

Also, I added a bunch of test coverage. We know now that the path finding actually works.